# cmd/api/admin-api/docs/swagger.yaml
basePath: /
definitions:
  apiresp.ApiResponse:
    properties:
      data: {}
      errCode:
        type: integer
      errDlt:
        type: string
      errMsg:
        type: string
    type: object
  github_com_openimsdk_chat_pkg_common_apistruct.AdminLoginResp:
    properties:
      adminAccount:
        type: string
      adminToken:
        type: string
      adminUserID:
        type: string
      faceURL:
        type: string
      imToken:
        type: string
      imUserID:
        type: string
      level:
        type: integer
      nickname:
        type: string
    type: object
...
  github_com_openimsdk_chat_pkg_protocol_admin.AddAdminAccountReq:
    properties:
      account:
        type: string
      faceURL:
        type: string
      nickname:
        type: string
      password:
        type: string
    type: object
  github_com_openimsdk_chat_pkg_protocol_admin.AddAppletReq:
    properties:
      appID:
        type: string
      createTime:
        type: integer
      icon:
        type: string
      id:
        type: string
      md5:
        type: string
      name:
        type: string
      priority:
        type: integer
      size:
        type: integer
      status:
        type: integer
      url:
        type: string
      version:
        type: string
    type: object
...
  github_com_openimsdk_chat_pkg_protocol_admin.AdminUpdateInfoReq:
    properties:
      account:
        $ref: '#/definitions/wrapperspb.StringValue'
      faceURL:
        $ref: '#/definitions/wrapperspb.StringValue'
      level:
        $ref: '#/definitions/wrapperspb.Int32Value'
      nickname:
        $ref: '#/definitions/wrapperspb.StringValue'
      password:
        $ref: '#/definitions/wrapperspb.StringValue'
    type: object
  github_com_openimsdk_chat_pkg_protocol_admin.ApplicationVersion:
    properties:
      createTime:
        type: integer
      force:
        type: boolean
      hot:
        type: boolean
      id:
        type: string
      latest:
        type: boolean
      platform:
        type: string
      text:
        type: string
      url:
        type: string
      version:
        type: string
    type: object
  github_com_openimsdk_chat_pkg_protocol_admin.BlockUserInfo:
    properties:
      account:
        type: string
      areaCode:
        type: string
      createTime:
        type: integer
      email:
        type: string
      faceURL:
        type: string
      gender:
        type: integer
      nickname:
        type: string
      opUserID:
        type: string
      phoneNumber:
        type: string
      reason:
        type: string
      userID:
        type: string
    type: object
  github_com_openimsdk_chat_pkg_protocol_admin.BlockUserReq:
    properties:
      reason:
        type: string
      userID:
        type: string
    type: object
...
  github_com_openimsdk_chat_pkg_protocol_admin.DelAppletReq:
    properties:
      appletIds:
        items:
          type: string
        type: array
    type: object
  github_com_openimsdk_chat_pkg_protocol_admin.DelClientConfigReq:
    properties:
      keys:
        items:
          type: string
        type: array
    type: object
...
  github_com_openimsdk_chat_pkg_protocol_admin.GetClientConfigReq:
    type: object
  github_com_openimsdk_chat_pkg_protocol_admin.GetClientConfigResp:
    properties:
      config:
        additionalProperties:
          type: string
        type: object
    type: object
...
  github_com_openimsdk_chat_pkg_protocol_admin.LoginReq:
    properties:
      account:
        type: string
      password:
        type: string
    type: object
  github_com_openimsdk_chat_pkg_protocol_admin.PageApplicationVersionReq:
    properties:
      pagination:
        $ref: '#/definitions/sdkws.RequestPagination'
      platform:
        items:
          type: string
        type: array
    type: object
  github_com_openimsdk_chat_pkg_protocol_admin.PageApplicationVersionResp:
    properties:
      total:
        type: integer
      versions:
        items:
          $ref: '#/definitions/github_com_openimsdk_chat_pkg_protocol_admin.ApplicationVersion'
        type: array
    type: object
  github_com_openimsdk_chat_pkg_protocol_admin.SearchAppletReq:
    properties:
      keyword:
        type: string
      pagination:
        $ref: '#/definitions/sdkws.RequestPagination'
    type: object
  github_com_openimsdk_chat_pkg_protocol_admin.SearchAppletResp:
    properties:
      applets:
        items:
          $ref: '#/definitions/github_com_openimsdk_chat_pkg_protocol_common.AppletInfo'
        type: array
      total:
        type: integer
    type: object
  github_com_openimsdk_chat_pkg_protocol_admin.SearchBlockUserReq:
    properties:
      keyword:
        type: string
      pagination:
        $ref: '#/definitions/sdkws.RequestPagination'
    type: object
  github_com_openimsdk_chat_pkg_protocol_admin.SearchBlockUserResp:
    properties:
      total:
        type: integer
      users:
        items:
          $ref: '#/definitions/github_com_openimsdk_chat_pkg_protocol_admin.BlockUserInfo'
        type: array
    type: object
...
  github_com_openimsdk_chat_pkg_protocol_admin.SetClientConfigReq:
    properties:
      config:
        additionalProperties:
          type: string
        type: object
    type: object
  github_com_openimsdk_chat_pkg_protocol_admin.UnblockUserReq:
    properties:
      userIDs:
        items:
          type: string
        type: array
    type: object
  github_com_openimsdk_chat_pkg_protocol_admin.UpdateAppletReq:
    properties:
      appID:
        $ref: '#/definitions/wrapperspb.StringValue'
      createTime:
        $ref: '#/definitions/wrapperspb.Int64Value'
      icon:
        $ref: '#/definitions/wrapperspb.StringValue'
      id:
        type: string
      md5:
        $ref: '#/definitions/wrapperspb.StringValue'
      name:
        $ref: '#/definitions/wrapperspb.StringValue'
      priority:
        $ref: '#/definitions/wrapperspb.UInt32Value'
      size:
        $ref: '#/definitions/wrapperspb.Int64Value'
      status:
        $ref: '#/definitions/wrapperspb.UInt32Value'
      url:
        $ref: '#/definitions/wrapperspb.StringValue'
      version:
        $ref: '#/definitions/wrapperspb.StringValue'
    type: object
...
  github_com_openimsdk_chat_pkg_protocol_common.AppletInfo:
    properties:
      appID:
        type: string
      createTime:
        type: integer
      icon:
        type: string
      id:
        type: string
      md5:
        type: string
      name:
        type: string
      priority:
        type: integer
      size:
        type: integer
      status:
        type: integer
      url:
        type: string
      version:
        type: string
    type: object
...
  sdkws.RequestPagination:
    properties:
      pageNumber:
        type: integer
      showNumber:
        type: integer
    type: object
...
  wrapperspb.Int32Value:
    properties:
      value:
        description: The int32 value.
        type: integer
    type: object
  wrapperspb.Int64Value:
    properties:
      value:
        description: The int64 value.
        type: integer
    type: object
  wrapperspb.StringValue:
    properties:
      value:
        description: The string value.
        type: string
    type: object
  wrapperspb.UInt32Value:
    properties:
      value:
        description: The uint32 value.
        type: integer
    type: object
host: localhost:10009
info:
  contact:
    email: ilovecoding@foxmail.com
    name: 飞宏IM技术支持
    url: https://github.com/liony823/chat
  description: 飞宏IM后台管理接口
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: 飞宏IM后台管理接口
  version: 1.0.0
paths:
  /account/add_admin:
    post:
      consumes:
      - application/json
      description: 创建新的管理员账户
      operationId: addAdminAccount
      parameters:
      - description: 管理员账户信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_openimsdk_chat_pkg_protocol_admin.AddAdminAccountReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apiresp.ApiResponse'
      summary: 添加管理员账户
      tags:
      - account
...
  /account/login:
    post:
      consumes:
      - application/json
      description: 管理员账户登录
      operationId: login
      parameters:
      - description: 登录信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_openimsdk_chat_pkg_protocol_admin.LoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/apiresp.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_openimsdk_chat_pkg_common_apistruct.AdminLoginResp'
              type: object
      summary: 管理员登录
      tags:
      - account
  /account/update:
    post:
      consumes:
      - application/json
      description: 更新管理员账户信息
      operationId: updateInfo
      parameters:
      - description: 管理员信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_openimsdk_chat_pkg_protocol_admin.AdminUpdateInfoReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apiresp.ApiResponse'
      summary: 更新管理员信息
      tags:
      - account
  /applet/add:
    post:
      consumes:
      - application/json
      description: 添加新的小程序配置
      operationId: addApplet
      parameters:
      - description: 小程序信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_openimsdk_chat_pkg_protocol_admin.AddAppletReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apiresp.ApiResponse'
      summary: 添加小程序
      tags:
      - applet
  /applet/del:
    post:
      consumes:
      - application/json
      description: 删除指定的小程序
      operationId: delApplet
      parameters:
      - description: 要删除的小程序ID
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_openimsdk_chat_pkg_protocol_admin.DelAppletReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apiresp.ApiResponse'
      summary: 删除小程序
      tags:
      - applet
  /applet/search:
    post:
      consumes:
      - application/json
      description: 根据条件搜索小程序列表
      operationId: searchApplet
      parameters:
      - description: 搜索条件
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_openimsdk_chat_pkg_protocol_admin.SearchAppletReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/apiresp.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_openimsdk_chat_pkg_protocol_admin.SearchAppletResp'
              type: object
      summary: 搜索小程序
      tags:
      - applet
  /applet/update:
    post:
      consumes:
      - application/json
      description: 更新小程序配置信息
      operationId: updateApplet
      parameters:
      - description: 更新的小程序信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_openimsdk_chat_pkg_protocol_admin.UpdateAppletReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apiresp.ApiResponse'
      summary: 更新小程序
      tags:
      - applet
...
  /application/page_versions:
    post:
      consumes:
      - application/json
      description: 分页获取应用程序版本列表
      operationId: pageVersions
      parameters:
      - description: 分页查询条件
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_openimsdk_chat_pkg_protocol_admin.PageApplicationVersionReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/apiresp.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_openimsdk_chat_pkg_protocol_admin.PageApplicationVersionResp'
              type: object
      summary: 分页查询应用版本
      tags:
      - application
...
  /block/add:
    post:
      consumes:
      - application/json
      description: 封禁指定用户账号
      operationId: addBlockUser
      parameters:
      - description: 要封禁的用户信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_openimsdk_chat_pkg_protocol_admin.BlockUserReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apiresp.ApiResponse'
      summary: 封禁用户
      tags:
      - block
  /block/del:
    post:
      consumes:
      - application/json
      description: 解除用户账号封禁
      operationId: unblockUser
      parameters:
      - description: 要解封的用户信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_openimsdk_chat_pkg_protocol_admin.UnblockUserReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apiresp.ApiResponse'
      summary: 解封用户
      tags:
      - block
  /block/search:
    post:
      consumes:
      - application/json
      description: 查询被封禁的用户列表
      operationId: searchBlockUser
      parameters:
      - description: 搜索条件
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_openimsdk_chat_pkg_protocol_admin.SearchBlockUserReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/apiresp.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_openimsdk_chat_pkg_protocol_admin.SearchBlockUserResp'
              type: object
      summary: 搜索被封禁用户
      tags:
      - block
  /client_config/del:
    post:
      consumes:
      - application/json
      description: 删除指定的客户端配置
      operationId: delClientConfig
      parameters:
      - description: 要删除的配置信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_openimsdk_chat_pkg_protocol_admin.DelClientConfigReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apiresp.ApiResponse'
      summary: 删除客户端配置
      tags:
      - client_config
  /client_config/get:
    post:
      consumes:
      - application/json
      description: 获取客户端初始化配置信息
      operationId: getClientConfig
      parameters:
      - description: 查询条件
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_openimsdk_chat_pkg_protocol_admin.GetClientConfigReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/apiresp.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_openimsdk_chat_pkg_protocol_admin.GetClientConfigResp'
              type: object
      summary: 获取客户端配置
      tags:
      - client_config
  /client_config/set:
    post:
      consumes:
      - application/json
      description: 设置客户端初始化配置信息
      operationId: setClientConfig
      parameters:
      - description: 客户端配置信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_openimsdk_chat_pkg_protocol_admin.SetClientConfigReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apiresp.ApiResponse'
      summary: 设置客户端配置
      tags:
      - client_config
...
securityDefinitions:
  BasicAuth:
    type: basic
  BearerAuth:
    description: '"Type ''Bearer'' followed by a space and JWT token"'
    in: header
    name: token
    type: apiKey
swagger: "2.0"